Python alter. (#1869)
### What problem does this PR solve?

Add python interface for alter statement.
Add pytest for alter statement.
Fix several bug in alter statement.
Refactor some restart test with python decorator.

Issue link:#[Link the issue here]

### Type of change

- [x] Bug Fix (non-breaking change which fixes an issue)
- [x] New Feature (non-breaking change which adds functionality)
- [x] Test cases

+++ src/storage/meta/entry/table_index_entry.cpp
@@ -372,26 +372,28 @@ Status TableIndexEntry::CreateIndexDo(BaseTableRef *table_ref, HashMap<SegmentID
     return Status::OK();
 }
 
-void TableIndexEntry::Cleanup(CleanupInfoTracer *info_tracer) {
+void TableIndexEntry::Cleanup(CleanupInfoTracer *info_tracer, bool dropped) {
     if (this->deleted_) {
         return;
     }
     std::unique_lock w_lock(rw_locker_);
     for (auto &[segment_id, segment_index_entry] : index_by_segment_) {
-        segment_index_entry->Cleanup(info_tracer);
+        segment_index_entry->Cleanup(info_tracer, dropped);
     }
 
-    LOG_DEBUG(fmt::format("Cleaning up dir: {}", *index_dir_));
+    if (dropped) {
+        LOG_DEBUG(fmt::format("Cleaning up dir: {}", *index_dir_));
 
-    LocalFileSystem fs;
-    String absolute_index_dir = fmt::format("{}/{}", InfinityContext::instance().config()->DataDir(), *index_dir_);
-    if (!fs.Exists(absolute_index_dir)) {
-        return;
-    }
-    fs.DeleteDirectory(absolute_index_dir);
-    LOG_DEBUG(fmt::format("Cleaned dir: {}", absolute_index_dir));
-    if (info_tracer != nullptr) {
-        info_tracer->AddCleanupInfo(std::move(absolute_index_dir));
+        LocalFileSystem fs;
+        String absolute_index_dir = fmt::format("{}/{}", InfinityContext::instance().config()->DataDir(), *index_dir_);
+        if (!fs.Exists(absolute_index_dir)) {
+            return;
+        }
+        fs.DeleteDirectory(absolute_index_dir);
+        LOG_DEBUG(fmt::format("Cleaned dir: {}", absolute_index_dir));
+        if (info_tracer != nullptr) {
+            info_tracer->AddCleanupInfo(std::move(absolute_index_dir));
+        }
     }
 }
